+++ src/StaticCollection.php
@@ -1,0 +1,275 @@
+<?php
+
+namespace olvlvl\ComposerAttributeCollector;
+
+use RuntimeException;
+use Throwable;
+
+use function array_map;
+
+/**
+ * An attribute collection that uses embedded data to provide instances.
+ *
+ * @internal
+ */
+final class StaticCollection implements Collection
+{
+    /**
+     * @param array<class-string, array<array{ string, class-string }>> $targetClasses
+     *     Where _key_ is an attribute class and _value_ an array of arrays
+     *     where 0 is the serialized attribute arguments and 1 is a target class.
+     * @param array<class-string, array<array{ string, class-string, non-empty-string }>> $targetMethods
+     *     Where _key_ is an attribute class and _value_ an array of arrays
+     *     where 0 is the serialized attribute arguments, 1 is a target class, and 2 is the target method.
+     * @param array<class-string, array<array{ string, class-string, non-empty-string }>> $targetProperties
+     *     Where _key_ is an attribute class and _value_ an array of arrays
+     *     where 0 is the serialized attribute arguments, 1 is a target class, and 2 is the target property.
+     * @param array<class-string, array<array{ string, class-string, non-empty-string, non-empty-string }>> $targetParameters
+     *     Where _key_ is an attribute class and _value_ an array of arrays
+     *     where 0 is the serialized attribute arguments, 1 is a target class, 2 is the target method, and 3 is the target parameter.
+     */
+    public function __construct(
+        private array $targetClasses,
+        private array $targetMethods,
+        private array $targetProperties,
+        private array $targetParameters,
+    ) {
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function findTargetClasses(string $attribute): array
+    {
+        return array_map(
+            fn(array $t) => self::createClassAttribute($attribute, ...$t),
+            $this->targetClasses[$attribute] ?? [],
+        );
+    }
+
+    /**
+     * @template T of object
+     *
+     * @param class-string<T> $attribute
+     * @param string $arguments The serialized arguments
+     * @param class-string $class
+     *
+     * @return TargetClass<T>
+     */
+    private static function createClassAttribute(string $attribute, string $arguments, string $class): object
+    {
+        try {
+            $a = new $attribute(...unserialize($arguments));
+            return new TargetClass($a, $class);
+        } catch (Throwable $e) {
+            throw new RuntimeException(
+                "An error occurred while instantiating attribute $attribute on class $class",
+                previous: $e,
+            );
+        }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function findTargetMethods(string $attribute): array
+    {
+        return array_map(
+            fn(array $t) => self::createMethodAttribute($attribute, ...$t),
+            $this->targetMethods[$attribute] ?? [],
+        );
+    }
+
+    /**
+     * @template T of object
+     *
+     * @param class-string<T> $attribute
+     * @param string $arguments The serialized arguments
+     * @param class-string $class
+     * @param non-empty-string $method
+     *
+     * @return TargetMethod<T>
+     */
+    private static function createMethodAttribute(
+        string $attribute,
+        string $arguments,
+        string $class,
+        string $method,
+    ): object {
+        try {
+            $a = new $attribute(...unserialize($arguments));
+            return new TargetMethod($a, $class, $method);
+        } catch (Throwable $e) {
+            throw new RuntimeException(
+                "An error occurred while instantiating attribute $attribute on method $class::$method",
+                previous: $e,
+            );
+        }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function findTargetParameters(string $attribute): array
+    {
+        return array_map(
+            fn(array $t) => self::createParameterAttribute($attribute, ...$t),
+            $this->targetParameters[$attribute] ?? [],
+        );
+    }
+
+    /**
+     * @template T of object
+     *
+     * @param class-string<T> $attribute
+     * @param string $arguments The serialized arguments
+     * @param class-string $class
+     * @param non-empty-string $method
+     * @param non-empty-string $parameter
+     *
+     * @return TargetParameter<T>
+     */
+    private static function createParameterAttribute(
+        string $attribute,
+        string $arguments,
+        string $class,
+        string $method,
+        string $parameter,
+    ): object {
+        try {
+            $a = new $attribute(...unserialize($arguments));
+            return new TargetParameter($a, $class, $method, $parameter);
+        } catch (Throwable $e) {
+            throw new RuntimeException(
+                "An error occurred while instantiating attribute $attribute on parameter $class::$method($parameter)",
+                previous: $e,
+            );
+        }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function findTargetProperties(string $attribute): array
+    {
+        return array_map(
+            fn(array $t) => self::createPropertyAttribute($attribute, ...$t),
+            $this->targetProperties[$attribute] ?? [],
+        );
+    }
+
+    /**
+     * @template T of object
+     *
+     * @param class-string<T> $attribute
+     * @param string $arguments The serialized arguments
+     * @param class-string $class
+     * @param non-empty-string $property
+     *
+     * @return TargetProperty<T>
+     */
+    private static function createPropertyAttribute(
+        string $attribute,
+        string $arguments,
+        string $class,
+        string $property,
+    ): object {
+        try {
+            $a = new $attribute(...unserialize($arguments));
+            return new TargetProperty($a, $class, $property);
+        } catch (Throwable $e) {
+            throw new RuntimeException(
+                "An error occurred while instantiating attribute $attribute on property $class::$property",
+                previous: $e,
+            );
+        }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function filterTargetClasses(callable $predicate): array
+    {
+        $ar = [];
+
+        foreach ($this->targetClasses as $attribute => $references) {
+            foreach ($references as [$arguments, $class]) {
+                if ($predicate($attribute, $class)) {
+                    $ar[] = self::createClassAttribute($attribute, $arguments, $class);
+                }
+            }
+        }
+
+        return $ar;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function filterTargetMethods(callable $predicate): array
+    {
+        $ar = [];
+
+        foreach ($this->targetMethods as $attribute => $references) {
+            foreach ($references as [$arguments, $class, $method]) {
+                if ($predicate($attribute, $class, $method)) {
+                    $ar[] = self::createMethodAttribute(
+                        $attribute,
+                        $arguments,
+                        $class,
+                        $method,
+                    );
+                }
+            }
+        }
+
+        return $ar;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function filterTargetParameters(callable $predicate): array
+    {
+        $ar = [];
+
+        foreach ($this->targetParameters as $attribute => $references) {
+            foreach ($references as [$arguments, $class, $method, $parameter]) {
+                if ($predicate($attribute, $class, $method, $parameter)) {
+                    $ar[] = self::createParameterAttribute(
+                        $attribute,
+                        $arguments,
+                        $class,
+                        $method,
+                        $parameter,
+                    );
+                }
+            }
+        }
+
+        return $ar;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function filterTargetProperties(callable $predicate): array
+    {
+        $ar = [];
+
+        foreach ($this->targetProperties as $attribute => $references) {
+            foreach ($references as [$arguments, $class, $property]) {
+                if ($predicate($attribute, $class, $property)) {
+                    $ar[] = self::createPropertyAttribute(
+                        $attribute,
+                        $arguments,
+                        $class,
+                        $property,
+                    );
+                }
+            }
+        }
+
+        return $ar;
+    }
+}