+++ tests/TestAbstract.php
@@ -1,12 +1,5 @@
 <?php
 
-/*
- * (c) Olivier Laviale <olivier.laviale@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace tests\olvlvl\ComposerAttributeCollector;
 
 use Acme\Attribute\ActiveRecord\Boolean;
@@ -27,13 +20,13 @@ use Acme\Attribute\Subscribe;
 use Acme\PSR4\Presentation\ArticleController;
 use Acme81\Attribute\ParameterA;
 use Acme81\Attribute\ParameterB;
+use Acme85\PSR4\SampleWithClosure;
 use olvlvl\ComposerAttributeCollector\Attributes;
 use olvlvl\ComposerAttributeCollector\Config;
 use olvlvl\ComposerAttributeCollector\TargetClass;
 use olvlvl\ComposerAttributeCollector\TargetMethod;
 use olvlvl\ComposerAttributeCollector\TargetParameter;
 use olvlvl\ComposerAttributeCollector\TargetProperty;
-use PhpParser\Node\Param;
 use PHPUnit\Framework\TestCase;
 use ReflectionException;
 
@@ -44,12 +37,19 @@ use function usort;
 
 use const PHP_VERSION_ID;
 
-abstract class CollectorTestAbstract extends TestCase
+abstract class TestAbstract extends TestCase
 {
     /**
      * @var array<class-string, bool>
      */
     private static array $initialized = [];
+
+    /**
+     * @return Config::STRATEGY_*
+     */
+    abstract protected static function getStrategy(): string;
+
+    abstract protected static function dump(Config $config): void;
 
     /**
      * @throws ReflectionException
@@ -62,7 +62,7 @@ abstract class CollectorTestAbstract extends TestCase
             return;
         }
 
-        $config = self::makeConfig();
+        $config = static::makeConfig();
 
         static::dump($config);
 
@@ -73,20 +73,23 @@ abstract class CollectorTestAbstract extends TestCase
         self::$initialized[get_called_class()] = true;
     }
 
-    abstract protected static function dump(Config $config): void;
-
-    private static function makeConfig(): Config
+    protected static function makeConfig(): Config
     {
         $cwd = getcwd();
         assert(is_string($cwd));
         $vendorDir = __DIR__ . '/sandbox';
         $filepath = "$vendorDir/attributes.php";
+        $strategy = static::getStrategy();
         $exclude = [
             "$cwd/tests/Acme/PSR4/IncompatibleSignature.php",
         ];
 
         if (PHP_VERSION_ID < 80100) {
             $exclude[] = "$cwd/tests/Acme81";
+        }
+
+        if (PHP_VERSION_ID < 80500 || $strategy != Config::STRATEGY_REFERENCE) {
+            $exclude[] = "$cwd/tests/Acme85";
         }
 
         return new Config(
@@ -98,6 +101,7 @@ abstract class CollectorTestAbstract extends TestCase
             exclude: $exclude,
             useCache: false,
             isDebug: false,
+            strategy: $strategy,
         );
     }
 
@@ -393,6 +397,27 @@ abstract class CollectorTestAbstract extends TestCase
         );
 
         $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * @requires PHP >= 8.5
+     */
+    public function testClosureInConstant(): void
+    {
+        if (static::getStrategy() != Config::STRATEGY_REFERENCE) {
+            $this->markTestSkipped();
+        }
+
+        $actual = Attributes::filterTargetProperties(
+            Attributes::predicateForAttributeInstanceOf(\Acme85\Attribute\WithClosure::class),
+        );
+
+        $this->assertCount(1, $actual);
+
+        /** @var \Acme85\Attribute\WithClosure $attribute */
+        $attribute = $actual[0]->attribute;
+        $this->assertInstanceOf(\Acme85\Attribute\WithClosure::class, $attribute);
+        $this->assertEquals(SampleWithClosure::MAGIC_STRING, ($attribute->closure)());
     }
 
     public function testFilterTargetParameters(): void